+++ AccountSystemLogic/src/main/java/za/ac/nwu/ac/logic/flow/ModifyMemberAccountFlow.java
@@ -3,5 +3,5 @@ package za.ac.nwu.ac.logic.flow;
 import za.ac.nwu.ac.domain.dto.MemberAccountDto;
 
 public interface ModifyMemberAccountFlow {
-    MemberAccountDto updateMemberAccount(Long amount, Long memberId, Long accountTypeId);
+    MemberAccountDto updateMemberAccount(Integer amount, Long memberId, Long accountTypeId);
 }
